refactor: consolidate content into base client
