+++ packages/catalog/src/registry/content/client-rpc.ts
@@ -31,9 +31,19 @@ export class RpcClient extends Context.Service<RpcClient>()("RpcClient", {
 
 // Client tick atom
 export const clientTickAtomContents = `import type { TickEvent } from "@repo/domain/Rpc";
-import { Effect, Stream } from "effect";
-import { runtime } from "../atom";
+import { Effect, Layer, Stream } from "effect";
+import { DevTools } from "effect/unstable/devtools";
+import { Atom } from "effect/unstable/reactivity";
 import { RpcClient } from "../rpc-client";
+
+const ENABLE_DEVTOOLS = import.meta.env.VITE_ENABLE_DEVTOOLS === "true";
+
+const RpcLayer = Layer.mergeAll(
+  RpcClient.layer,
+  ENABLE_DEVTOOLS ? DevTools.layer() : Layer.empty,
+);
+
+const runtime = Atom.runtime(RpcLayer);
 
 export const tickAtom = runtime.fn(
   ({ abort = false }: { readonly abort?: boolean }) =>
@@ -121,20 +131,4 @@ export const RpcCard = () => {
     </div>
   );
 };
-`;
-
-// Client atom.ts with RpcClient layer
-export const clientAtomWithRpcContents = `import { Layer } from "effect";
-import { DevTools } from "effect/unstable/devtools";
-import { Atom } from "effect/unstable/reactivity";
-import { RpcClient } from "./rpc-client";
-
-const ENABLE_DEVTOOLS = import.meta.env.VITE_ENABLE_DEVTOOLS === "true";
-
-const MainLayer = Layer.mergeAll(
-  RpcClient.layer,
-  ENABLE_DEVTOOLS ? DevTools.layer() : Layer.empty,
-);
-
-export const runtime = Atom.runtime(MainLayer);
 `;
